Refactor behaviour y más sorpresas (#32)
* Creación de party (bastante barebones)

se crea una party en la base de datos pero de momento nada más

* Creación de parties más pulida

* Ya puede uno unirse a parties

* Fixes del merge con main

* Salida de party y borrado de parties

* Inspeccionar habitaciones en compañía

* Moverse por el mapa en compañía

* Status de los compañeros de party

* Status de la party + inicio de giveItem

* no guardé

* Squashed commit of the following:

commit 9056cd1804a7db09db7aae7bf40f291f2a2fbeff
Author: Miguel Mur Cortés <mmur@ucm.es>
Date:   Sat Mar 26 21:44:56 2022 +0100

    Movidos textos de feedback al usuario del sistema de inventario.

    También he cambiado que haga falta pasarles un string a que ahora le tengas que pasar un item

* Envío de objetos a compañeros de party

* Más fixes del merge

* Más fixes de merges wiii

* Arreglado lo de la referencia del player

* Algún avance (aún no compila)

* ya se inician las batallas multijugador (no compila)

No hace nada más, lo he probado comentando todo lo que no está hecho de momento

* Medio funciona la batalla???????

* Más avances en la batalla yee haw

* Modo comunismo

* Qué \n Qué sucede con el multiplayer

* Multiplayer 100% mejorado (sigue sin funcionar 100%)


* It just works?????????


* Fix un jugador

* Empieza el refactor

* Algunos fixes

* Refactor terminado?

* El ataque de los clones

* fixes

* El retorno de enemy y muerte general

* Fixes

* cambio float a int

* ints

* oncreate en enemy

* JSONSystem más eficiente

* Carga del mapa a JSONSystem

También arreglados comandos

+++ MMOTFG_Bot/MMOTFG_Bot/Battle/Attacks/aScaled.cs
@@ -8,20 +8,20 @@ namespace MMOTFG_Bot
 {
     class aScaled : Attack
     {
-        public aScaled(string name_, float power_, float mpCost_) : base(name_, power_, mpCost_) { }
+        public aScaled(string name_, float power_, int mpCost_) : base(name_, power_, mpCost_) { }
 
         [JsonConverter(typeof(StringEnumConverter))]
         public StatName statToScale;
 
-        public float fixedDamage;
+        public int fixedDamage;
 
-        public override float GetDamage()
+        public override int GetDamage()
         {
             //ataque de daño fijo
             if (fixedDamage != 0) return fixedDamage;
 
             //ataque que escala con el otro stat
-            return user.GetStat(statToScale) * power;
+            return (int)(user.GetStat(statToScale) * power);
         }
     }
 }
